refactor: extract interactive_terminal.py and mattermost_display.py
- Move run_interactive + helpers from agent.py to interactive_terminal.py (170 lines)
- Move ConversationDisplay from mattermost.py to mattermost_display.py (383 lines)
- agent.py: 1080 → 920 lines, mattermost.py: 1340 → 973 lines
